refactor(interactive_shell): rename host_ports.py to host_contracts.py (#5367)
