+++ tools/interactive_shell/shared/host_contracts.py
